+++ Skin/Interface/AddOns/Blizzard_ActionBarController/Blizzard_ActionBarController.lua
@@ -576,16 +576,13 @@ function private.FrameXML.Blizzard_ActionBarController()
 
 
     ----====####$$$$%%%%%$$$$####====----
-    --           MainMenuBar           --
+    --           MainActionBar           --
     ----====####$$$$%%%%%$$$$####====----
     if not private.disabled.mainmenubar then
         if private.isRetail then
-            local MainMenuBar = _G.MainMenuBar
-            MainMenuBar.BorderArt:SetAlpha(0)
-            -- Removed in 10.2.0
-            -- MainMenuBar.Background:SetAlpha(0)
-            MainMenuBar.EndCaps:SetAlpha(0)
-
+            local MainActionBar = _G.MainActionBar
+            MainActionBar.BorderArt:SetAlpha(0)
+            MainActionBar.EndCaps:SetAlpha(0)
         else
             _G.hooksecurefunc("MainMenuTrackingBar_Configure", Hook.MainMenuTrackingBar_Configure)
 
@@ -666,8 +663,9 @@ function private.FrameXML.Blizzard_ActionBarController()
     ----====####$$$$%%%%%$$$$####====----
     if not private.disabled.mainmenubar and private.isRetail then
         Util.Mixin(_G.StatusTrackingBarManager, Hook.StatusTrackingManagerMixin)
-        Skin.StatusTrackingBarContainerTemplate(_G.MainStatusTrackingBarContainer)
-        Skin.StatusTrackingBarContainerTemplate(_G.SecondaryStatusTrackingBarContainer)
+        -- FIXLATER -- this is now a Mixin, so we need to hook the :New method to apply the mixin to new bars
+        -- Skin.StatusTrackingBarContainerTemplate(_G.MainStatusTrackingBarContainer)
+        -- Skin.StatusTrackingBarContainerTemplate(_G.SecondaryStatusTrackingBarContainer)
     end
 
 
@@ -704,6 +702,8 @@ function private.FrameXML.Blizzard_ActionBarController()
     ----====####$$$$%%%%%$$$$####====----
     --- FIXBETA - why is this here?
     --- Bartender4 workaround!
+    --- If Bartender4 is loaded, the end caps and border art are not hidden properly.
+    --- This is fixed in newer versions of Bartender4, but keeping this for older versions.
     local LeftEndCap = _G.MainActionBar.EndCaps.LeftEndCap
     local RightEndCap = _G.MainActionBar.EndCaps.RightEndCap
     local BoarderArt = _G.MainActionBar.BorderArt