+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-﻿## [11.2.7.0] ##
+﻿## [11.2.7.2] ##
+### Fixed ###
+  * chg: moved from NUM_CHAT_WINDOWS to Constants.ChatFrameConstants.MaxChatWindows
+
+## [11.2.7.1] ##
+### Fixed ###
+  * chg: MainMenuBar replaced by MainActionBar
+  * fix: channel editbox info when chatType is CHANNEL
+  * chg: ScrollBox.view.poolCollection replaced by ScrollUpdate hooks
+
+## [11.2.7.0] ##
 ### Fixed ###
   * chg: renamed toc file to support _optional versions..
   * chore: toc bump